+++ app/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.6.1'
     implementation 'com.squareup.retrofit2:converter-gson:2.6.1'
     implementation 'com.google.firebase:firebase-messaging:19.0.1'
+    implementation 'com.github.jakebonk:ChipView:1.0.1'
