+++ app/src/main/res/layout/dashboard_activity.xml
@@ -49,22 +49,8 @@
             android:layout_gravity="start"
             app:headerLayout="@layout/nav_header"
             app:menu="@menu/drawer_menu" />
-
-
-
-
-
-
-
-
-
-
-
-
-
     </androidx.drawerlayout.widget.DrawerLayout>
     <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:onClick="ShowPopup"
         android:id="@+id/askQ"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
